apps/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.11.0
pydantic-settings==2.7.0